requirements.txt
streamlit==1.10.0
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.4
streamlit-embedcode==0.1.2
streamlit-extras==0.2.7
streamlit-faker==0.0.2
streamlit-image-coordinates==0.1.3
streamlit-keyup==0.2.0
streamlit-option-menu==0.3.2
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
scikit-learn==1.1.1
scikit-learn-intelex==2021.20220215.102710
pmdarima==2.0.3
conda==22.9.0
conda-build==3.21.8
conda-verify==3.4.2
statsmodels==0.13.2
pipreqs==0.4.11
hijri-converter==2.2.4
anaconda-navigator==2.3.2
pandas==1.5.3
plotly==5.6.0
numpy==1.21.5
altair==4.1.0

